Handled order table

+++ database/migrations/2023_05_15_081515_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->integer('total_price');
+            $table->double('total_price');
             $table->string('address');
             $table->string('payment_method');
             $table->double('money_received');
